removed unnecessary hooks and simplified some function call
