+++ src/apps/settings/src/app/components/TopBar.tsx
@@ -62,7 +62,6 @@ export const TopBar: FC<TopBarProps> = ({
             color="text.primary"
             noWrap
             minWidth="fit-content"
-            sx={{ flexGrow: 1, width: "100%" }}
           >
             {title}
           </Typography>
