+++ src/shell/services/instance.ts
@@ -30,7 +30,6 @@ import {
   GroupItem,
   Redirects,
   RedirectRequest,
-  WebFont,
 } from "./types";
 import { batchApiRequests } from "../../utility/batchApiRequests";
 
@@ -66,8 +65,6 @@ export const instanceApi = createApi({
     "WorkflowStatusLabels",
     "Groups",
     "Redirects",
-    "WebFonts",
-    "LegacyHeadTags",
   ],
   endpoints: (builder) => ({
     // https://www.zesty.io/docs/instances/api-reference/content/models/items/publishings/#Get-All-Item-Publishings
@@ -508,7 +505,6 @@ export const instanceApi = createApi({
     getLegacyHeadTags: builder.query<LegacyHeader[], void>({
       query: () => `/web/headers`,
       transformResponse: getResponseData,
-      providesTags: ["LegacyHeadTags"],
     }),
     // https://www.zesty.io/docs/instances/api-reference/env/settings/#Get-All-Settings
     getInstanceSettings: builder.query<InstanceSetting[], void>({
@@ -902,19 +898,6 @@ export const instanceApi = createApi({
       }),
       invalidatesTags: ["Redirects"],
     }),
-    getWebFonts: builder.query<WebFont[], void>({
-      async queryFn(args, _queryApi, _extraOptions, fetchWithBQ) {
-        try {
-          const res: any = await fetchWithBQ({
-            url: "https://www.googleapis.com/webfonts/v1/webfonts?key=AIzaSyAWGIDvIvAF69UQZR9vIOtz88s55deGs8Y",
-          });
-          return { data: res?.data?.items };
-        } catch (error) {
-          return { error };
-        }
-      },
-      providesTags: ["WebFonts"],
-    }),
   }),
 });
 
@@ -982,6 +965,5 @@ export const {
   useDeleteRedirectMutation,
   useLazySearchContentQuery,
   useUpdateRedirectMutation,
-  useGetWebFontsQuery,
   useUpdateHeadTagsMutation,
 } = instanceApi;
